+++ lab6/index.html
@@ -15,6 +15,9 @@
         <div id="ball-box">
             <div id="ball"></div>
         </div>
+        <button id="buttonForDumTitlesKillMe">
+            What could THAT be for?
+        </button>
     </body>
     <script src="script.js"></script>
 </html>
