--- lab6/task5.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="stylesheet" href="styles.css">
-        <title>Lab6</title>
-    </head>
-    <body>
-        <button id="buttonForDumTitlesKillMe">
-            What could THAT be for?
-        </button>
-    </body>
-    <script src="task5.js"></script>
-</html>
